⬆️ Update zaw plugin to latest version
Why is this change needed?
The zaw plugin submodule needs to be updated to incorporate
the latest bug fixes and improvements from upstream.

How does it address the issue?
This updates the zaw submodule pointer from commit 2033f9b
to commit 2404177, bringing in the latest changes from the
zaw repository.